Enhancement to Safari automation
Soliciting feedback on change.  The idea is to save the manual step of selecting the file to save the json results.  Instead, the manual tester can accept the Safari defailt (~/Downloads/Accessibility Result.json) and the script will rename the file in the local directory based on our naming convention.  Requires the user supply the Jira ticket number on the command line.  If this change looks good, I'll update the README.md.